Add User-Agent header to track SDK usage (#170)
* Move staticClient tests into directory.

* Split tests into separate files.

* Use scope.isDone() instead of nock.isDone().
scope.isDone checks that a single mock request is done.
nock.isDone checks that all mock requests are done.

* Enforce User-Agent header.
Format is "commerce-sdk@version;"

* Clarify test descriptions.

* Clean duplicate code to improve readability.

* Fix: ensure default user agent is replaced.
make-fetch-happen default must be replaced using lower case header.

* Add user agent and strip header tests.

* Use require w/o import to avoid TypeScript errors.

* Add User-Agent integration test.
