+++ packages/core/tsconfig.json
@@ -1,7 +1,7 @@
 {
     "extends": "../../base.tsconfig.json",
     "compilerOptions": {
-        "outDir": "./dist",
+        "outDir": "./dist"
     },
     // NOTE: Everything except what you want compiled into JS should be in the exclusions.
     "exclude": [
